Implement terminal reloading

+++ src/ts/terminal/index.ts
@@ -7,12 +7,23 @@ import { LoginUser } from "$ts/server/user/auth";
 import type { UserDaemon } from "$ts/server/user/daemon";
 import { ElevationLevel, type ElevationData } from "$types/elevation";
 import type { DirectoryReadReturn } from "$types/fs";
-import type { Arguments } from "$types/terminal";
+import type { ArcTermConfiguration, Arguments } from "$types/terminal";
 import ansiEscapes from "ansi-escapes";
 import type { Terminal } from "xterm";
 import { Readline } from "./readline/readline";
-import { BOLD, BRBLACK, BRBLUE, BRGREEN, BRRED, BRYELLOW, RESET, TerminalCommandStore } from "./store";
+import {
+  BOLD,
+  BRBLACK,
+  BRBLUE,
+  BRGREEN,
+  BRRED,
+  BRYELLOW,
+  DefaultArcTermConfiguration,
+  RESET,
+  TerminalCommandStore,
+} from "./store";
 import { ArcTermVariables } from "./var";
+import { arrayToText, textToBlob } from "$ts/fs/convert";
 
 export class ArcTerminal extends Process {
   path: string;
@@ -24,6 +35,7 @@ export class ArcTerminal extends Process {
   daemon: UserDaemon | undefined;
   ansiEscapes = ansiEscapes;
   lastCommandErrored = false;
+  config: ArcTermConfiguration = DefaultArcTermConfiguration;
 
   constructor(handler: ProcessHandler, pid: number, parentPid: number, term: Terminal, path?: string) {
     super(handler, pid, parentPid);
@@ -38,6 +50,7 @@ export class ArcTerminal extends Process {
     if (!this.term) return this.killSelf();
 
     const rl = await this.handler.spawn<Readline>(Readline, undefined, this.pid, this);
+    await this.readConfig();
 
     this.term.loadAddon(rl!);
     this.rl = rl;
@@ -46,14 +59,16 @@ export class ArcTerminal extends Process {
   }
 
   async readline() {
-    const username = this.env.get("currentuser") || "Stranger";
-    const color = this.lastCommandErrored ? BRRED : RESET;
-    const line = await this.rl?.read(`${BRGREEN}${username}${RESET}: ${BRGREEN}${this.path} ${color}$${RESET} `);
+    if (this._disposed) return;
+
+    const line = await this.rl?.read(this.var?.replace(this.config.prompt || "$")!);
 
     await this.processLine(line);
   }
 
   async processLine(text: string | undefined) {
+    if (this._disposed) return;
+
     this.lastCommandErrored = false;
 
     if (!text) return this.readline();
@@ -85,6 +100,8 @@ export class ArcTerminal extends Process {
   }
 
   join(path?: string) {
+    if (this._disposed) return "";
+
     if (!path) return this.path;
     if (path.includes(":/")) return path;
 
@@ -92,50 +109,74 @@ export class ArcTerminal extends Process {
   }
 
   async readDir(path?: string) {
+    if (this._disposed) return;
+
     return await this.fs.readDir(this.join(path));
   }
 
   async createDirectory(path: string) {
+    if (this._disposed) return;
+
     return await this.fs.createDirectory(this.join(path));
   }
 
   async writeFile(path: string, data: Blob) {
+    if (this._disposed) return;
+
     return await this.fs.writeFile(this.join(path), data);
   }
 
   async tree(path: string) {
+    if (this._disposed) return;
+
     return await this.fs.tree(this.join(path));
   }
 
   async copyItem(source: string, destination: string) {
+    if (this._disposed) return;
+
     return await this.fs.copyItem(this.join(source), this.join(destination));
   }
 
   async moveItem(source: string, destination: string) {
+    if (this._disposed) return;
+
     return await this.fs.moveItem(this.join(source), this.join(destination));
   }
 
   async readFile(path: string) {
+    if (this._disposed) return;
+
     return await this.fs.readFile(this.join(path));
   }
 
   async deleteItem(path: string) {
+    if (this._disposed) return;
+
     return await this.fs.deleteItem(this.join(path));
   }
 
   async Error(message: string, prefix = "Error") {
+    if (this._disposed) return;
+
     this.rl?.println(`${BRRED}${prefix}${RESET}: ${message}`);
   }
 
   async Warning(message: string, prefix = "Warning") {
+    if (this._disposed) return;
+
     this.rl?.println(`${BRYELLOW}${prefix}${RESET}: ${message}`);
   }
 
   async Info(message: string, prefix = "Info") {
+    if (this._disposed) return;
+
     this.rl?.println(`${BRBLUE}${prefix}${RESET}: ${message}`);
   }
 
   async changeDirectory(path: string) {
+    if (this._disposed) return;
+
     try {
       const drive = this.fs.getDriveByPath(path);
 
@@ -162,6 +203,8 @@ export class ArcTerminal extends Process {
   }
 
   parseFlags(args: string): [Arguments, string] {
+    if (this._disposed) return [{}, ""];
+
     const regex = /(?:--(?<nl>[a-z\-]+)(?:="(?<vl>.*?)"|(?:=(?<vs>.*?)(?: |$))|)|-(?<ns>[a-zA-Z]))/gm; //--name=?value
     const matches: RegExpMatchArray[] = [];
 
@@ -195,6 +238,7 @@ export class ArcTerminal extends Process {
   }
 
   async elevate(data: ElevationData) {
+    if (this._disposed) return false;
     const color = data.level == ElevationLevel.low ? BRGREEN : data.level === ElevationLevel.medium ? BRYELLOW : BRRED;
     const pref = this.daemon?.preferences();
 
@@ -252,5 +296,32 @@ export class ArcTerminal extends Process {
     await this.daemon?.discontinueToken(token);
 
     return true;
+  }
+
+  async readConfig() {
+    if (this._disposed) return;
+    try {
+      const contents = await this.fs.readFile("U:/arcterm.conf");
+
+      if (!contents) throw "";
+
+      const json = JSON.parse(arrayToText(contents));
+
+      this.config = json as ArcTermConfiguration;
+    } catch {
+      await this.writeConfig();
+    }
+  }
+
+  async writeConfig() {
+    if (this._disposed) return;
+
+    await this.fs.writeFile("U:/arcterm.conf", textToBlob(JSON.stringify(this.config, null, 2)));
+  }
+
+  async reload() {
+    await this.rl?.dispose();
+    await this.killSelf();
+    await this.handler.spawn(ArcTerminal, undefined, this.parentPid, this.term, this.path);
   }
 }
